refactor: Redis Service에서 Component 로 변경
